requirements_with_airflow.txt
alembic==1.4.2
apache-airflow==1.10.11
apispec==1.3.3
argcomplete==1.12.0
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
cached-property==1.5.1
cattrs==1.0.0
cffi==1.14.1
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.34
cryptography==3.0
defusedxml==0.6.0
dill==0.3.2
dnspython==2.0.0
docutils==0.16
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.14.1
gunicorn==19.10.0
importlib-metadata==1.7.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.5.1
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
natsort==7.0.1
pendulum==1.4.4
prison==0.1.3
psutil==5.7.2
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyrsistent==0.16.0
python-daemon==2.2.4
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytzdata==2020.1
setproctitle==1.1.10
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
typing==3.7.4.3
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
Werkzeug==0.16.1
WTForms==2.3.1
zipp==3.1.0
zope.deprecation==4.4.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
loguru==0.5.3
numpy==1.19.4
oauth2client
pandas==1.1.4
protobuf==3.13.0
psycopg2-binary==2.8.6
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive
pyspark==3.0.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.11
